SDL: bugfix random hang
Sometimes the SDL application would hang at random times.
This happened when the time to render a frame was longer than the
frequency wanted (eg 16.666 ms).
The delay computed would be negative and converted to a big `uint32_t`
when given to the `SDL_Delay()` function, resulting in a total hang of
the app.
We now only call `SDL_Delay()` if rendering the frame took less time
than the wanted frequency.