Lib: bugfix opcode `FX0A` wait for input
Previously this would only look for a pressed key and return it, thus
not waiting if a key was pressed before the opcode execution.
Now, once a key is released after the opcode execution, the program will
continue. Even tho this key was pressed before the call which seems odd
to me but apparently it is the expected behavior.